dom/tests/mochitest/whatwg/test_postMessage_override.html

<!DOCTYPE html>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=postMessage
-->
<head>
  <title>postMessage override test</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=postMessage">Mozilla Bug 387706</a>
<p id="display"></p>
<div id="content" style="display: none"></div>

<iframe src="http://example.org:8000/tests/dom/tests/mochitest/whatwg/postMessage_override_helper.html"></iframe>


<pre id="test">
<script class="testbody" type="application/javascript">
/** Test for Bug 387706 **/

SimpleTest.waitForExplicitFinish();

async function run() {
  function ping(win, msg) {
    return new Promise(resolve => {
      window.addEventListener("message", evt => {
        is(evt.source, win, "Response should come from the correct window");
        is(evt.data, msg, "Should get the correct response");
        resolve();
      }, { once: true });

      win.postMessage(msg, "http://example.org:8000");
    });
  }

  await ping(frames[0], "PASS 1");
  await ping(frames[0], "PASS 2");
  SimpleTest.finish();
}

window.addEventListener("load", run);
</script>
</pre>
</body>
</html>
